+++ Enderecos.py
@@ -1,9 +1,14 @@
-STATUSMARIO = 0x72B481
-MOEDAS = 0x72C227
-YOSHICOINS = 0x72C888
+import sys
+if sys.platform == 'linux':
+    STATUSMARIO = 0x569f4699
+    MOEDAS = 0x569f543f
+    YOSHICOINS = 0x569f5aa2
+else:
+    STATUSMARIO = 0x72B481
+    MOEDAS = 0x72C227
+    YOSHICOINS = 0x72C888
 
 MARIO_PEQUENO = 0
 MARIO_GRANDE = 1
 MARIO_PENINHA = 2
 MARIO_FLOR_DE_FOGO = 3
-
